+++ priv/template/web/controllers/page_controller.ex
@@ -1,6 +1,8 @@
 defmodule <%= application_module %>.PageController do
   use Phoenix.Controller
 
+  plug :action
+
   def index(conn, _params) do
     render conn, "index"
   end