+++ priv/template/web/router.ex
@@ -1,6 +1,15 @@
 defmodule <%= application_module %>.Router do
   use Phoenix.Router
 
-  get "/", <%= application_module %>.PageController, :index, as: :pages
+  scope "/" do
+    # Use the default browser stack.
+    pipe_through :browser
 
+    get "/", <%= application_module %>.PageController, :index, as: :pages
+  end
+
+  # Other scopes may use custom stacks.
+  # scope "/api" do
+  #   pipe_through :api
+  # end
 end