+++ test/phoenix/controller/render_test.exs
@@ -6,7 +6,8 @@ Code.require_file "views/render_view.exs", __DIR__
 defmodule MyApp.AssignController do
   use Phoenix.Controller
 
-  before_action :assign_value
+  plug :assign_value
+  plug :action
 
   def assign_value(conn, _) do
     assign(conn, :my_assign, "assign_plug")
@@ -35,7 +36,8 @@ end
 defmodule MyApp.ImplicitRenderController do
   use Phoenix.Controller
 
-  after_action :render
+  plug :action
+  plug :render
 
   def index(conn, _params) do
     conn
@@ -46,6 +48,8 @@ end
 
 defmodule MyApp.RenderController do
   use Phoenix.Controller
+
+  plug :action
 
   def show(conn, _params) do
     render conn, "show", []