+++ test/phoenix/router/console_formatter_test.exs
@@ -1,5 +1,5 @@
 defmodule Phoenix.Router.ConsoleFormatterTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
   alias Phoenix.Router.ConsoleFormatter
 
   defmodule RouterTestSingleRoutes do