+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Enhancements
   * [Router] Named helpers are now automatically generated for every route based on the controller name
   * [Router] Named helpers have been optimized to do as little work as possible at runtime
+  * [Router] Support multiple pipelines at the router level
   * [Channels] The `phoenix.js` channel client now sends a configurable heartbeat every 30s to maintain connections
 
 * Deprecations
@@ -13,10 +14,11 @@
 
 * Backwards incompatible changes
   * [Controller] Remove default, inject aliases: `Flash`, `JSON`
-  * [Controller] Controllers now use `before_action` and `after_action` to customize the plug stack
+  * [Controller] Controllers now require `plug :action` to be explicitly invoked
   * [Router] `*path` identifiers in routers are now returned as a list
   * [Router] Named helpers are now defined in a explicit module nested to your router. For example, if your router is named `MyApp.Router`, the named helpers will be available at `MyApp.Router.Helpers`
   * [Router] `session_secret` configuration is deprecated in favor of `secret_key_base`
+  * [Router] Plugs can now only be defined inside pipelines. All routers now need to explicitly declare which pipeline they want to use.
 
 
 ## v0.4.1 (2014-09-08)